Change width of fpu_state_t for x86_64.
gcc (and probably other compilers) don't like working with 16-bit
types and will zero-extend where needed. Save some overhead and
just store the state as a 32-bit type.

+++ arch/x86_64/fpu/fpu.h
@@ -11,7 +11,7 @@
 #define __arch_fpu_h__
 #include "common.h"
 
-typedef uint16_t fpu_state_t;
+typedef uint32_t fpu_state_t;
 
 #define FPU_MASK_EXCPS    0x1F80
